Update AP_Mission.cpp: cos()/sin() -> cosf()/sinf()
to conform with DO_NOT_USE_DOUBLE_MATHS()

+++ libraries/AP_Mission/AP_Mission.cpp
@@ -1704,8 +1704,8 @@ bool AP_Mission::advance_current_nav_cmd(uint16_t starting_index)
                                         tmp.direction =+ 36000;
                                     }
                                     rel_distance = sqrtf((rel_lat*rel_lat)+(rel_lng*rel_lng));
-                                    tmp.lat = (int32_t)(cos((float)tmp.direction/100.0*DEG_TO_RAD)*rel_distance);
-                                    tmp.lng = (int32_t)(sin((float)tmp.direction/100.0*DEG_TO_RAD)*rel_distance)/cmd.content.location.longitude_scale();
+                                    tmp.lat = (int32_t)(cosf((float)tmp.direction/100.0*DEG_TO_RAD)*rel_distance);
+                                    tmp.lng = (int32_t)(sinf((float)tmp.direction/100.0*DEG_TO_RAD)*rel_distance)/cmd.content.location.longitude_scale();
                                     cmd.content.location.lat = _start_loc.lat + tmp.lat;
                                     cmd.content.location.lng = _start_loc.lng + tmp.lng;
                                 }
